storage: implement CocoonFs
Implement of CocoonFs, a secure special purpose filesystem format intended
to be used in untrusted environments, like e.g. in a TEE setting.

For a specification of the filesystem format, refer to the file
at storage/src/fs/cocoonfs/cocoonfs-format.md

RustDoc comments for the internal APIs is still work in progress,
but publish the code now already in order to have something to discuss
about.

Signed-off-by: Nicolai Stange <nstange@suse.de>