cocoonfs: decode ImageLayout only after the static header has been read
The ImageLayout decoding involves a validation of the various
filesystem configuration paramters. In order to avoid
spurious/misleading errors, it should be delayed to after the upcoming
CRC-32 integrity check on the full static image header has passed. In
preparation to this, move the ImageLayout decoding to after the full
static image header has been read in.

Signed-off-by: Nicolai Stange <nstange@suse.de>