cocoonfs: let the exported mkfs primitives take the image size in bytes
Currently both, the CocoonFsMkFsFuture and the
CocoonFsWriteMkFsInfoHeader, expect the desired image size parameter in
units of Allocation Blocks. However, AllocBlockCount is not public,
and the concept of Allocation Blocks is also a concept internal to the
CocoonFS format. Let both filesystem creation related primitives take
the desired filesystem image size parameter in the more natural unit
of bytes.

Signed-off-by: Nicolai Stange <nstange@suse.de>