utils-async: allow for passing aux data to the wrapped Future's poll()
In preparation of providing access to a RNG by passing a reference to
the CocoonFS' respective NvFsFuture's poll() instead of assuming
ownership of it, introduce support for passing an auxiliary data
argument to the BroadcastFuture's wrapped inner Future.

To this end, provide a new trait, BroadcastedFuture, defining this
extended poll() interface and pass the auxiliary poll data from the
subscriptions' poll() to the that BroadcastedFuture::poll().

Signed-off-by: Nicolai Stange <nstange@suse.de>