crypto/symcipher: drop debug_assert!() on lengths in inner loop
The debug_assert!() on the IO slice iterators' lengths in the symcipher
transformation's inner loop, i.e. in transform_next_blocks!() and
transform_next_blocks_in_place!() respectively, make the whole
operation quadratic in the input length (or rather on the number of
buffers in each of the two IO slice iterators. This is an actual
problem, even for debug builds, in certain workloads. Drop those
debug_assert!()s.

Signed-off-by: Nicolai Stange <nstange@suse.de>