cocoonfs: optimize storage allocation algorithm(s)
The storage allocation algorithms, the ones for for power-of-two sized
blocks as well as for small extents to be more specific, try hard to
find the best possible placement to reduce overall fragmentation
pressure. They do so by means of a global search over _all_ of the
allocation bitmap for the least bad option.

Especially if the filesystem image, hence the allocation bitmap, is
large, this becomes a performance issue.

Restrict the search to a fixed, constant placement optimization search
distance of 16 bitmap words, starting from the first found free
candiate location. With the minimum possible Allocation Block size of
128B, this effectively limits the search range to
16 * BitmapWord::BITS * 128B == 128kB of storage.

For short-lived allocations, i.e. those made for the journal, it
doesn't make much sense to optimize the placement at all, as these
don't contribute to the fragmentation in practice. Introduce a new
argument to the allocation primitives, 'optimize_placement', and set
it to false as appropriate when doing journal allocations.

Lastly, find_free_block() always starts the search at the beginning
of the allocation bitmap. However, if find_free_block() is invoked in
a loop, it is better to resume the search at the point the last free
block had been found in a previous iteration -- it is known that there
is no other free block available before that position and this avoids
a superfluous reexamination. Make find_free_block() accept an Option
for indicating where to start the search. If None, it will start at
the beginning. Once the end of the allocation bitmap is reached in the
search, it will wrap around and proceed up to the initial starting
point.

Signed-off-by: Nicolai Stange <nstange@suse.de>