utils-async: fix miri stacked borrows violation in generic_arc::Arc
By taking a reference to the outer ArcInner in the course of deriving
a pointer to the wrapped value in Arc::as_ptr(), the latter carries a
stacked borrow to the former and miri complains. Same for
Weak::as_ptr().

Fix this by deriving the pointers to the wrapped value by means of
a raw reference op.

Signed-off-by: Nicolai Stange <nstange@suse.de>