storage: rename "NvChip" to "NvBlkDev"
The trait "NvChip" defines a generic interface to a block device.
The naming is confusing though -- it's not really descriptive by
itself and also, misleading in the context context of the TPM 2.0
reference implementation, where the term "NV chip" is used to refer
to the NV state data blob.

Rename "NvChip" to "NvBlkDev". Mechanically mass-rename all related
identifiers accordingly.

There is no change in functionality.

Suggested-by: Stefano Garzarella <sgarzare@redhat.com>
Suggested-by: Oliver Steffen <osteffen@redhat.com>
Signed-off-by: Nicolai Stange <nstange@suse.de>