cocoonfs: make StartReadSequenceFuture state private
The public CocoonFsStartReadSequenceFuture is a direct wrapper around
the internal enum StartReadSequenceFuture.

Currently, StartReadSequenceFuture is not exposed to extern, but
this will change: CocoonFsStartReadSequenceFuture is about to get
dropped and replaced by StartReadSequenceFuture.

With StartReadSequenceFuture being an enum currently, all its internal
state would get exposed to the outside then. In order to avoid this,
represent the state by means of the new StartReadSequenceFutureState
and make StartReadSequenceFuture a struct containing a member thereof.

There is no change in functionality.

Signed-off-by: Nicolai Stange <nstange@suse.de>