storage: make NvFs::WriteInodeFut::poll() to always return the data
The  NvFs::WriteInodeFut is supposed to take ownership of the input
data buffer for the duration of its operation and to eventually
return it back upon completion. However, right now, it is not
expected to always return it back unconditionally -- a stale
transaction as indicated by NvFsError::Retry is an example where it
would not return the data back.

This is an API misdesign though, because the caller can hardly conduct
a retry if it doesn't have the data buffer.

Make NvFs::WriteInodeFut::poll() to always return the input data
buffer back upon completion.

Signed-off-by: Nicolai Stange <nstange@suse.de>